Modified Craftdata Patching to allow for dynamic changes so as not to be fully dependedent on load order.
ALSO paves the way for making CC2 be able to have a reload changes option where a person could change the recipes out of game and use a options menu toggle to load said changes.    Could also allow cc2 do a second pass as a world save is loading to redo the patches which would fix it so people dont have to load game then exit then load again . but that will need to be done on cc2's side.