More dynamic vanilla TechType check
Now the vanilla TechType checker is more smart on what is a vanilla TechType. Previously the check was merely comparing the TechType integer number against the highest enum value; this works, however, it will require us to manually increase the value each time a game update introduces a new item.