Interfaces/FishHandler Purge
Convert Handlers to Public Static
Delete All Interfaces
Deleted FishHanlder and related files.
Deleted Handler class.
Deleted BioReactorHandler due to is only patching at patch time AND since that dictionary is super simple to modify without being in smlhelper.
Added a couple simple methods that will make it easier to keep the same code for both games.
Removed the Sound systems that were using Sound Channels that were marked obsolete.
Set the storyGoals list creation to only happen in SN1 as that field doesnt exist in bz.
Deleted PDAEncylopediaHandler and PDALogHandler and moved them into PDA Handler as there was only 1 method in both of them.

Added in a check to the PingTypePatcher to verify that a Vanilla PingType does not already exist with the name we are trying to add.

Accidentally ran my formatter over the solution.....   sowwy