Change the way the ModOptions menu's are built to allow more modder capabilities.
Created a RemoveItem(string Id) to remove options
Set both AddItem and RemoveItem to return true if they succedded or False if they did not.

Made the BuildModOptions a virtual method that actually does the building of the menu that modders can override to do it themselves. To allow this I have passed the panel, modsTabIndex and currently added Mods from that ModOptions into it.

Made it so that the options no longer get wiped every time the options panel closes allowing you to add all your options in the ModOptions constructor and then just let SMLHelper build the menu without having to override it every time the menu opens.