Make CraftData.GetTechType check PrefabIdentifier over TechTag.
This 1 patch allows mod techtypes to be properly loaded by the game even if the cache is messed up or deleted from my testing!

Any mods that still rely on saving the INT value of the TechType instead of the name will still have trouble but that is a problem those mods need to fix as it should only happen in their own loading systems.