ModPrefabBuilder beginnings
Moved the ModPrefab dictionaries to the ModPrefabCache where they belong.
Created a ModPrefabBuilder to attach my Extensions to.
Removed the PrefabHandler in place of an Extension method that returns a ModPrefabBuilder
Added Extensions to the ModPrefabBuilder to handle all the different parts of the Assets that used to be handled by the Handlers.
Kept adding more extensions.
Still want to add even more eventually but this is a good time to stop for now.
Fixed a couple incorrect documentations that I found.