Make Custom Batteries actually work.
Added extra NeedsPatching check to CraftDataPatcher that also repatches if the game is about to wipe the cache.

Fixed bug where SMLHelper sometimes failed to figure out what mod added a Enum Value.

Added in the custom batteries code that moved Batteries and PowerCells into new tabs.

Sealed off EasyBatery and EasyPowerCell to just be pretty much prefab templates.

Made a bunch of stuf public that was internal from when it was still in CB but now needs to be public for CB to still do its text plugins.