Modpocalypse 2 Electric Boogaloo
Asset Interfaces.
Refactor Custom Batteries stuff to be more inline with SMLHelpers systems.
Make ModPrefabCache actually a cache instead of just a thing that deletes prefabs after 30 seconds.

Bye Bye Abstract assets and PrefabHandler

Move all patching to awake.  Will probably break something... but they should not need to be done after mods are loaded if we have the patching done right.
And if we don't then it needs to be fixed instead of hidden by delaying the patches.

Fixed the Reflection Helper saying the Battery tabs that we added were being added by unity instead of smlhelper.

Removed automatic calling of the sync version of GetGameObject unless the person inherits IModPrefabSync and sets the action.