More efficient localization handling
Previously, the localization dictionary for folder specifically was getting reset each time the language was changing, since it's all IO, it is worse to re-read localization files everytime.

Now we load them all to our customLines dictionary at once, which makes them consistent with how the rest of the API works.